Trait dvm_compiler::disassembler::code::translator::Context

pub trait Context<'a> {
    fn pop_exp(&mut self) -> ExpLoc<'a>;

    fn last_exp(&self) -> Option<&ExpLoc<'a>>;

    fn pop2_exp(&mut self) -> (ExpLoc<'a>, ExpLoc<'a>);

    fn pop_exp_vec(&mut self, exp_count: usize) -> Vec<ExpLoc<'a>>;

    fn module_import(&self, module: &ModuleHandle) -> Option<Import<'a>>;

    fn extract_signature(
        &self, 
        type_params: Option<&SignatureIndex>
    ) -> Vec<FType<'a>>;

    fn local_var(&self, index: u8) -> Local<'a>;

    fn opcode_offset(&self) -> usize;

    fn pack_fields(&mut self, def: &StructDefinition) -> Vec<PackField<'a>>;

    fn translate_block(&mut self, block_size: usize) -> Vec<ExpLoc<'a>>;

    fn next_opcode(&mut self) -> Option<&Bytecode>;

    fn loc(&self, exp: Exp<'a>) -> ExpLoc<'a>;

    fn opcode_by_relative_offset(&self, offset: isize) -> &Bytecode;

    fn opcode_by_absolute_offset(&self, offset: usize) -> &Bytecode;

    fn end_offset(&self) -> usize;

    fn remaining_code(&self) -> &[Bytecode];

    fn err(&self) -> Exp<'a>;
}

Transaction context. Provides functions for bytecode transactions.

Required methods

fn pop_exp(&mut self) -> ExpLoc<'a>

Removes the last element from a expression list and returns it, or [Exp::Non] if it is empty.

fn last_exp(&self) -> Option<&ExpLoc<'a>>

Returns reference to the last element of a expression list or [None] if it is empty.

fn pop2_exp(&mut self) -> (ExpLoc<'a>, ExpLoc<'a>)

Removes the two last elements from a expression list and returns it, or [Exp::Non] if it is empty.

fn pop_exp_vec(&mut self, exp_count: usize) -> Vec<ExpLoc<'a>>

Removes the exp_count last elements from a expression list and returns it.

fn module_import(&self, module: &ModuleHandle) -> Option<Import<'a>>

Returns module Import by its handle reference.

fn extract_signature(
&self,
type_params: Option<&SignatureIndex>
) -> Vec<FType<'a>>

Extracts signature by its index.

fn local_var(&self, index: u8) -> Local<'a>

Returns local variable by its index.

fn opcode_offset(&self) -> usize

Returns current bytecode offset.

fn pack_fields(&mut self, def: &StructDefinition) -> Vec<PackField<'a>>

Returns struct fields by its definition.

fn translate_block(&mut self, block_size: usize) -> Vec<ExpLoc<'a>>

Translates next block_size bytecode instructions and returns it.

fn next_opcode(&mut self) -> Option<&Bytecode>

Returns next bytecode instruction and updates bytecode iterator state.

fn loc(&self, exp: Exp<'a>) -> ExpLoc<'a>

Wraps the given expression at the current location.

fn opcode_by_relative_offset(&self, offset: isize) -> &Bytecode

Returns the bytecode instruction by relative index.

fn opcode_by_absolute_offset(&self, offset: usize) -> &Bytecode

Returns the bytecode instruction by absolute index.

fn end_offset(&self) -> usize

Returns the last bytecode offset of the current context.

fn remaining_code(&self) -> &[Bytecode]

Returns remaining bytecode instructions.

fn err(&self) -> Exp<'a>

Returns error expression.

Loading content...